double -> BigInteger, BigDecimal 로 교체하는 작업. 계산 결과의 정밀도를 높일 수 있다고 한다.

+++ app/src/main/java/kr/asv/loancalculator/Amortization.java
@@ -1,8 +1,10 @@
 package kr.asv.loancalculator;
 
+import java.math.BigInteger;
+
 interface Amortization
 {
 	void calculate(LoanCalculatorOptions options);
 	PaymentSchedules getSchedules();
-	double getSummaryInterest();
+	BigInteger getSummaryInterest();
 }
