+++ app/src/main/java/kr/asv/loancalculator/Simulator.java
@@ -1,5 +1,8 @@
 package kr.asv.loancalculator;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 /**
  * 진입점
  */
@@ -10,9 +13,9 @@ class Simulator
 		LoanCalculator calculator = new LoanCalculator();
 		
 		LoanCalculatorOptions options = calculator.getOptions();
-		options.setPrincipal(4700000.0);//원금
+		options.setPrincipal(BigInteger.valueOf(4700000));//원금
 		options.setAmortizationPeriod(36);//상환기간
-		options.setInterestRate(19.45);//이자율
+		options.setInterestRate(BigDecimal.valueOf(19.45));//이자율
 		options.setAmortizationMethod(LoanCalculator.AmortizationMethods.EQUAL_PRINCIPAL);
 		options.setDebug(true);
